requirements.txt
Flask==2.3.2
tensorflow==2.10.1
numpy
Pillow
flask-cors
